Refactor more
